@@ -22,18 +22,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.4",
-    "@angular/common": "^4.3.4",
-    "@angular/compiler": "^4.3.4",
-    "@angular/core": "^4.3.4",
+    "@angular/animations": "4.4.3",
+    "@angular/common": "4.4.3",
+    "@angular/compiler": "4.4.3",
+    "@angular/core": "4.4.3",
     "@angular/flex-layout": "2.0.0-beta.9",
-    "@angular/forms": "^4.3.4",
-    "@angular/http": "^4.3.4",
-    "@angular/material": "^2.0.0-beta.10",
-    "@angular/cdk": "^2.0.0-beta.10",
-    "@angular/platform-browser": "^4.3.4",
-    "@angular/platform-browser-dynamic": "^4.3.4",
-    "@angular/router": "^4.3.4",
+    "@angular/forms": "4.4.3",
+    "@angular/http": "4.4.3",
+    "@angular/material": "2.0.0-beta.11",
+    "@angular/cdk": "2.0.0-beta.11",
+    "@angular/platform-browser": "4.4.3",
+    "@angular/platform-browser-dynamic": "4.4.3",
+    "@angular/router": "4.4.3",
     "@nationallibraryofnorway/ngx-mime": "../dist/",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.4.1",
@@ -45,8 +45,8 @@
   },
   "devDependencies": {
     "@angular/cli": "1.3.0",
-    "@angular/compiler-cli": "^4.3.4",
-    "@angular/language-service": "^4.3.4",
+    "@angular/compiler-cli": "4.4.3",
+    "@angular/language-service": "4.4.3",
     "@types/chai": "3.5.2",
     "@types/chai-as-promised": "0.0.30",
     "@types/cucumber": "2.0.1",
