@@ -44,27 +44,27 @@
     "release": "sh scripts/release/release.sh"
   },
   "peerDependencies": {
-    "@angular/core": ">=4.3.0 <5.0.0 || >=4.0.0-beta <5.0.0",
-    "@angular/material": ">=2.0.0 <3.0.0 || >=2.0.0-beta.10 <3.0.0",
+    "@angular/core": ">=4.4.3 <5.0.0",
+    "@angular/material": ">=2.0.0 <3.0.0 || >=2.0.0-beta.11 <3.0.0",
     "@angular/cdk": ">=2.0.0 <3.0.0 || >=2.0.0-beta.8 <3.0.0",
     "@angular/flex-layout": ">=2.0.0 <3.0.0 || >=2.0.0-beta.9 <3.0.0",
     "openseadragon": ">=2.0.0 <3.0.0",
     "d3": ">=4.1.0 <5.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "4.3.5",
-    "@angular/cdk": "2.0.0-beta.10",
-    "@angular/common": "4.3.5",
-    "@angular/compiler": "4.3.5",
-    "@angular/compiler-cli": "4.3.5",
-    "@angular/core": "4.3.5",
+    "@angular/animations": "4.4.3",
+    "@angular/cdk": "2.0.0-beta.11",
+    "@angular/common": "4.4.3",
+    "@angular/compiler": "4.4.3",
+    "@angular/compiler-cli": "4.4.3",
+    "@angular/core": "4.4.3",
     "@angular/flex-layout": "2.0.0-beta.9",
-    "@angular/forms": "4.3.5",
-    "@angular/http": "4.3.5",
-    "@angular/material": "2.0.0-beta.10",
-    "@angular/platform-browser": "4.3.5",
-    "@angular/platform-browser-dynamic": "4.3.5",
-    "@angular/router": "4.3.5",
+    "@angular/forms": "4.4.3",
+    "@angular/http": "4.4.3",
+    "@angular/material": "2.0.0-beta.11",
+    "@angular/platform-browser": "4.4.3",
+    "@angular/platform-browser-dynamic": "4.4.3",
+    "@angular/router": "4.4.3",
     "hammerjs": "2.0.8",
     "openseadragon": "^2.2.1",
     "@types/jasmine": "2.5.54",
