@@ -1,19 +1,19 @@
-
-import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
+import { DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { By } from '@angular/platform-browser';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
 import { MdDialogRef } from '@angular/material';
 import { Observable } from 'rxjs/Observable';
 
-import { SharedModule } from './../shared/shared.module';
+import { SharedModule } from '../shared/shared.module';
 import { AttributionDialogComponent } from './attribution-dialog.component';
 import { AttributionDialogResizeService } from './attribution-dialog-resize.service';
-import { IiifManifestService } from './../core/iiif-manifest-service/iiif-manifest-service';
-import { MimeViewerIntl } from './../core/viewer-intl';
-import { Manifest } from './../core/models/manifest';
+import { IiifManifestService } from '../core/iiif-manifest-service/iiif-manifest-service';
+import { MimeViewerIntl } from '../core/viewer-intl';
+import { Manifest } from '../core/models/manifest';
+import { MimeDomHelper } from '../core/mime-dom-helper';
+import { FullscreenService } from '../core/fullscreen-service/fullscreen.service';
 
 describe('AttributionDialogComponent', () => {
   let component: AttributionDialogComponent;
@@ -30,6 +30,8 @@ describe('AttributionDialogComponent', () => {
       providers: [
         MimeViewerIntl,
         AttributionDialogResizeService,
+        MimeDomHelper,
+        FullscreenService,
         { provide: IiifManifestService, useClass: IiifManifestServiceStub },
         { provide: MdDialogRef, useClass: MdDialogRefMock },
       ]
