@@ -4,6 +4,7 @@ import { SharedModule } from './../shared/shared.module';
 import { AttributionDialogComponent } from './attribution-dialog.component';
 import { AttributionDialogService } from './attribution-dialog.service';
 import { AttributionDialogResizeService } from './attribution-dialog-resize.service';
+import { MimeDomHelper } from '../core/mime-dom-helper';
 
 @NgModule({
   imports: [
@@ -14,7 +15,8 @@ import { AttributionDialogResizeService } from './attribution-dialog-resize.serv
   ],
   providers: [
     AttributionDialogService,
-    AttributionDialogResizeService
+    AttributionDialogResizeService,
+    MimeDomHelper
   ],
   entryComponents: [
     AttributionDialogComponent
