@@ -1,9 +1,11 @@
 import { TestBed, inject } from '@angular/core/testing';
 
-import { SharedModule } from './../shared/shared.module';
+import { SharedModule } from '../shared/shared.module';
 import { AttributionDialogService } from './attribution-dialog.service';
-import { MimeResizeService } from './../core/mime-resize-service/mime-resize.service';
+import { MimeResizeService } from '../core/mime-resize-service/mime-resize.service';
 import { AttributionDialogResizeService } from './attribution-dialog-resize.service';
+import { MimeDomHelper } from '../core/mime-dom-helper';
+import { FullscreenService } from '../core/fullscreen-service/fullscreen.service';
 
 describe('AttributionDialogService', () => {
   beforeEach(() => {
@@ -14,6 +16,8 @@ describe('AttributionDialogService', () => {
       providers: [
         AttributionDialogService,
         MimeResizeService,
+        MimeDomHelper,
+        FullscreenService,
         AttributionDialogResizeService
       ]
     });
