@@ -1,5 +1,5 @@
 import { MimeDomHelper } from './../core/mime-dom-helper';
-import { ElementRef } from '@angular/core';
+import { ElementRef, Injectable } from '@angular/core';
 import { MdDialogConfig } from '@angular/material';
 
 import { Dimensions } from './../core/models/dimensions';
@@ -13,7 +13,7 @@ export class MobileContentsDialogConfigStrategy implements ContentsDialogConfigS
   public getConfig(elementRef: ElementRef): MdDialogConfig {
     return {
       hasBackdrop: false,
-      disableClose: true,
+      disableClose: false,
       width: '100%',
       height: '100%'
     };
@@ -23,12 +23,17 @@ export class MobileContentsDialogConfigStrategy implements ContentsDialogConfigS
 export class DesktopContentsDialogConfigStrategy implements ContentsDialogConfigStrategy {
   public static readonly dialogWidth = 350;
   public static readonly paddingRight = 20;
+  private mimeDomHelper: MimeDomHelper;
+
+  constructor(mimeDomHelper: MimeDomHelper) {
+    this.mimeDomHelper = mimeDomHelper;
+  }
 
   public getConfig(el: ElementRef): MdDialogConfig {
     const dimensions = this.getPosition(el);
     return {
       hasBackdrop: false,
-      disableClose: true,
+      disableClose: false,
       width: `${DesktopContentsDialogConfigStrategy.dialogWidth}px`,
       position: {
         top: dimensions.top + 'px',
@@ -38,7 +43,7 @@ export class DesktopContentsDialogConfigStrategy implements ContentsDialogConfig
   }
 
   private getPosition(el: ElementRef): Dimensions {
-    const dimensions = new MimeDomHelper().getBoundingClientRect(el);
+    const dimensions = this.mimeDomHelper.getBoundingClientRect(el);
     return new Dimensions({
       top: dimensions.top + 64,
       left: dimensions.right - DesktopContentsDialogConfigStrategy.dialogWidth - DesktopContentsDialogConfigStrategy.paddingRight
