@@ -1,9 +1,11 @@
 import { TestBed, inject } from '@angular/core/testing';
 
-import { SharedModule } from './../shared/shared.module';
+import { SharedModule } from '../shared/shared.module';
 import { ContentsDialogService } from './contents-dialog.service';
-import { MimeResizeService } from './../core/mime-resize-service/mime-resize.service';
+import { MimeResizeService } from '../core/mime-resize-service/mime-resize.service';
 import { ContentsDialogConfigStrategyFactory } from './contents-dialog-config-strategy-factory';
+import { MimeDomHelper } from '../core/mime-dom-helper';
+import { FullscreenService } from '../core/fullscreen-service/fullscreen.service';
 
 describe('ContentsDialogService', () => {
   beforeEach(() => {
@@ -14,6 +16,8 @@ describe('ContentsDialogService', () => {
       providers: [
         ContentsDialogService,
         MimeResizeService,
+        MimeDomHelper,
+        FullscreenService,
         ContentsDialogConfigStrategyFactory
       ]
     });
