@@ -1,6 +1,6 @@
 import { Observable } from 'rxjs/Observable';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
-import { Injectable, ElementRef } from '@angular/core';
+import { Injectable } from '@angular/core';
 
 @Injectable()
 export class FullscreenService {
