@@ -1,9 +1,11 @@
 import { FullscreenService } from './fullscreen-service/fullscreen.service';
 import { Dimensions } from './models/dimensions';
-import { ElementRef } from '@angular/core';
+import { ElementRef, Injectable } from '@angular/core';
 
+@Injectable()
 export class MimeDomHelper {
-  private fullscreen = new FullscreenService();
+
+  constructor(private fullscreen: FullscreenService) { }
 
   public getBoundingClientRect(el: ElementRef): Dimensions {
     try {
