@@ -2,8 +2,8 @@ import { Injectable, ElementRef } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
 
-import { MimeDomHelper } from './../mime-dom-helper';
-import { Dimensions } from './../models/dimensions';
+import { MimeDomHelper } from '../mime-dom-helper';
+import { Dimensions } from '../models/dimensions';
 
 @Injectable()
 export class MimeResizeService {
@@ -11,7 +11,7 @@ export class MimeResizeService {
   private resizeSubject: ReplaySubject<Dimensions> = new ReplaySubject();
   private dimensions = new Dimensions();
 
-  constructor() { }
+  constructor(private mimeDomHelper: MimeDomHelper) { }
 
   set el(el: ElementRef) {
     this._el = el;
@@ -26,7 +26,7 @@ export class MimeResizeService {
   }
 
   markForCheck() {
-    const dimensions = new MimeDomHelper().getBoundingClientRect(this.el);
+    const dimensions = this.mimeDomHelper.getBoundingClientRect(this.el);
 
     if (this.dimensions.bottom !== dimensions.bottom ||
       this.dimensions.height !== dimensions.height ||
