@@ -53,11 +53,11 @@ export class Metadata {
 }
 
 export class Sequence {
-  public id: string;
-  public type: string;
-  public label: string;
-  public viewingHint: string;
-  public canvases: Canvas[];
+  public id?: string;
+  public type?: string;
+  public label?: string;
+  public viewingHint?: string;
+  public canvases?: Canvas[];
 
   constructor(
     fields?: {
@@ -79,13 +79,13 @@ export class Sequence {
 }
 
 export class Canvas {
-  public id: string;
-  public type: string;
-  public label: string;
-  public thumbnail: string;
-  public height: number;
-  public width: number;
-  public images: Images[];
+  public id?: string;
+  public type?: string;
+  public label?: string;
+  public thumbnail?: string;
+  public height?: number;
+  public width?: number;
+  public images?: Images[];
 
   constructor(
     fields?: {
@@ -166,17 +166,17 @@ export class Resource {
 }
 
 export class Service {
-  public context: string;
-  public id: string;
-  public protocol: string;
-  public width: number;
-  public height: number;
-  public sizes: Size[];
-  public tiles: Tile[];
-  public profile: string;
-  public physicalScale: number;
-  public physicalUnits: string;
-  public service: Service;
+  public context?: string;
+  public id?: string;
+  public protocol?: string;
+  public width?: number;
+  public height?: number;
+  public sizes?: Size[];
+  public tiles?: Tile[];
+  public profile?: string;
+  public physicalScale?: number;
+  public physicalUnits?: string;
+  public service?: Service;
 
   constructor(
     fields?: {
