@@ -1,10 +1,10 @@
-import { CenterPoints } from './page-center-point';
+import { TileRects } from './tile-rects';
 
 describe('CenterPoints ', () => {
 
   it('should return closest index', () => {
     const array = [10, 20, 30, 40, 50, 60];
-    const centerPoints = new CenterPoints();
+    const centerPoints = new TileRects();
     centerPoints.add({ x: 10, y: 0 });
     centerPoints.add({ x: 20, y: 0 });
     centerPoints.add({ x: 30, y: 0 });
