@@ -6,6 +6,7 @@ import { CoreModule } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
 import { ContentsDialogModule } from './contents-dialog/contents-dialog.module';
 import { AttributionDialogModule } from './attribution-dialog/attribution-dialog.module';
+import { ContentSearchDialogModule } from './content-search-dialog/content-search-dialog.module';
 import { ViewerComponent } from './viewer/viewer.component';
 import { ViewerHeaderComponent } from './viewer/viewer-header/viewer-header.component';
 import { ViewerFooterComponent } from './viewer/viewer-footer/viewer-footer.component';
@@ -29,8 +30,8 @@ import 'd3';
     CoreModule,
     SharedModule,
     ContentsDialogModule,
-    BrowserAnimationsModule,
-    AttributionDialogModule
+    AttributionDialogModule,
+    ContentSearchDialogModule
   ],
   exports: [
     ViewerComponent
