@@ -1,6 +1,7 @@
 import { SpinnerService } from '../core/spinner-service/spinner.service';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { FormsModule }   from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
 import { MimeMaterialModule } from './mime-material.module';
@@ -9,12 +10,14 @@ import { MimeMaterialModule } from './mime-material.module';
   imports: [
     CommonModule,
     FlexLayoutModule,
-    MimeMaterialModule
+    MimeMaterialModule,
+    FormsModule
   ],
   exports: [
     CommonModule,
     FlexLayoutModule,
-    MimeMaterialModule
+    MimeMaterialModule,
+    FormsModule
   ],
   providers: [
     SpinnerService
