@@ -1,15 +1,17 @@
-import { CustomOptions } from '../../core/models/options-custom';
 import { Component, ChangeDetectionStrategy, OnInit, OnDestroy, ChangeDetectorRef, Input, Renderer2, ElementRef } from '@angular/core';
 import { MdDialog, MdDialogConfig, DialogPosition } from '@angular/material';
 import { ObservableMedia, MediaChange } from '@angular/flex-layout';
 import { trigger, state, style, animate, transition } from '@angular/animations';
 import { Subscription } from 'rxjs/Subscription';
 
+import { CustomOptions } from '../../core/models/options-custom';
 import { MimeViewerIntl } from './../../core/viewer-intl';
 import { Manifest } from './../../core/models/manifest';
 import { ContentsDialogComponent } from './../../contents-dialog/contents-dialog.component';
 import { ContentsDialogService } from './../../contents-dialog/contents-dialog.service';
+import { ContentSearchDialogService } from './../../content-search-dialog/content-search-dialog.service';
 import { MimeDomHelper } from '../../core/mime-dom-helper';
+import { IiifManifestService } from './../../core/iiif-manifest-service/iiif-manifest-service';
 import { FullscreenService } from './../../core/fullscreen-service/fullscreen.service';
 
 @Component({
@@ -41,13 +43,17 @@ import { FullscreenService } from './../../core/fullscreen-service/fullscreen.se
 export class ViewerHeaderComponent implements OnInit, OnDestroy {
   private subscriptions: Array<Subscription> = [];
   public state = 'show';
+  isContentSearchEnabled = false;
   isFullscreenEnabled = false;
 
   constructor(
     public intl: MimeViewerIntl,
     private changeDetectorRef: ChangeDetectorRef,
     private contentsDialogService: ContentsDialogService,
-    private fullscreenService: FullscreenService) { }
+    private contentSearchDialogService: ContentSearchDialogService,
+    private iiifManifestService: IiifManifestService,
+    private fullscreenService: FullscreenService,
+    private mimeDomHelper: MimeDomHelper) { }
 
   ngOnInit() {
     this.isFullscreenEnabled = this.fullscreenService.isEnabled();
@@ -55,6 +61,11 @@ export class ViewerHeaderComponent implements OnInit, OnDestroy {
     this.subscriptions.push(this.intl.changes.subscribe(() => this.changeDetectorRef.markForCheck()));
 
     this.subscriptions.push(this.fullscreenService.onChange.subscribe(() => {
+      this.changeDetectorRef.detectChanges();
+    }));
+
+    this.subscriptions.push(this.iiifManifestService.currentManifest.subscribe((manifest: Manifest) => {
+      this.isContentSearchEnabled = manifest.service ? true : false;
       this.changeDetectorRef.detectChanges();
     }));
 
@@ -66,12 +77,18 @@ export class ViewerHeaderComponent implements OnInit, OnDestroy {
     });
   }
 
-  public openContents() {
+  public toggleContents() {
+    this.contentSearchDialogService.close();
     this.contentsDialogService.toggle();
   }
 
+  public toggleSearch() {
+    this.contentsDialogService.close();
+    this.contentSearchDialogService.toggle();
+  }
+
   public toggleFullscreen(): void {
-    return new MimeDomHelper().toggleFullscreen();
+    return this.mimeDomHelper.toggleFullscreen();
   }
 
   public isInFullScreen(): boolean {
