@@ -11,15 +11,13 @@ import {
   SimpleChanges,
   ViewChild,
 } from '@angular/core';
-import { MdDialog } from '@angular/material';
 import { Subscription } from 'rxjs/Subscription';
 import { IiifManifestService } from '../core/iiif-manifest-service/iiif-manifest-service';
 import { ContentsDialogService } from '../contents-dialog/contents-dialog.service';
 import { AttributionDialogService } from '../attribution-dialog/attribution-dialog.service';
+import { ContentSearchDialogService } from '../content-search-dialog/content-search-dialog.service';
 import { MimeResizeService } from '../core/mime-resize-service/mime-resize.service';
 import { Manifest } from '../core/models/manifest';
-import { Options } from '../core/models/options';
-import { PageService } from '../core/page-service/page-service';
 import { ModeService } from '../core/mode-service/mode.service';
 import { ViewerMode } from '../core/models/viewer-mode';
 import { CustomOptions } from '../core/models/options-custom';
@@ -29,6 +27,8 @@ import { OsdToolbarComponent } from './osd-toolbar/osd-toolbar.component';
 import { ViewerService } from '../core/viewer-service/viewer.service';
 import { MimeViewerConfig } from '../core/mime-viewer-config';
 import { Dimensions } from '../core/models/dimensions';
+import { IiifContentSearchService } from './../core/iiif-content-search-service/iiif-content-search.service';
+import { SearchResult } from './../core/models/search-result';
 
 @Component({
   selector: 'mime-viewer',
@@ -38,9 +38,11 @@ import { Dimensions } from '../core/models/dimensions';
 })
 export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
   @Input() public manifestUri: string;
+  @Input() public q: string;
   @Input() public config: MimeViewerConfig = new MimeViewerConfig();
   private subscriptions: Array<Subscription> = [];
   private isCanvasPressed = false;
+  private currentManifest: Manifest;
 
   ViewerMode: typeof ViewerMode = ViewerMode;
 
@@ -54,14 +56,15 @@ export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
     private iiifManifestService: IiifManifestService,
     private contentsDialogService: ContentsDialogService,
     private attributionDialogService: AttributionDialogService,
+    private contentSearchDialogService: ContentSearchDialogService,
     private viewerService: ViewerService,
     private mimeService: MimeResizeService,
-    private dialog: MdDialog,
     private changeDetectorRef: ChangeDetectorRef,
-    private pageService: PageService,
-    private modeService: ModeService) {
+    private modeService: ModeService,
+    private iiifContentSearchService: IiifContentSearchService) {
     contentsDialogService.el = el;
     attributionDialogService.el = el;
+    contentSearchDialogService.el = el;
     mimeService.el = el;
   }
 
@@ -70,12 +73,23 @@ export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
     this.subscriptions.push(
       this.iiifManifestService.currentManifest
         .subscribe((manifest: Manifest) => {
+          this.currentManifest = manifest;
           this.cleanUp();
           this.viewerService.setUpViewer(manifest);
           if (this.config.attributionDialogEnabled && manifest.attribution) {
             this.attributionDialogService.open(this.config.attributionDialogHideTimeout);
           }
+
+          if (this.q) {
+            this.iiifContentSearchService.search(manifest, this.q);
+          }
         })
+    );
+
+    this.subscriptions.push(
+      this.iiifContentSearchService.onChange.subscribe((sr: SearchResult) => {
+        this.viewerService.highlight(sr);
+      })
     );
 
     this.subscriptions.push(
@@ -93,14 +107,29 @@ export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
   }
 
   ngOnChanges(changes: SimpleChanges): void {
+    let manifestUriIsChanged = false;
+    let qIsChanged = false;
+    if (changes['q']) {
+      const qChanges: SimpleChange = changes['q'];
+      if (!qChanges.isFirstChange() && qChanges.currentValue !== qChanges.firstChange) {
+        this.q = qChanges.currentValue;
+        qIsChanged = true;
+      }
+    }
     if (changes['manifestUri']) {
       const manifestUriChanges: SimpleChange = changes['manifestUri'];
       if (manifestUriChanges.currentValue !== manifestUriChanges.previousValue) {
         this.modeService.mode = this.config.initViewerMode;
         this.manifestUri = manifestUriChanges.currentValue;
-        this.cleanUp();
-        this.loadManifest();
+        manifestUriIsChanged = true;
       }
+    }
+
+    if (manifestUriIsChanged) {
+      this.cleanUp();
+      this.loadManifest();
+    } else if (qIsChanged) {
+      this.iiifContentSearchService.search(this.currentManifest, this.q);
     }
   }
 
@@ -108,6 +137,16 @@ export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
     this.subscriptions.forEach((subscription: Subscription) => {
       subscription.unsubscribe();
     });
+    this.iiifManifestService.destroy();
+    this.iiifContentSearchService.destroy();
+  }
+
+  // ChangeDetection fix
+  onModeChange() {
+    if (this.mode === ViewerMode.DASHBOARD) {
+      this.contentsDialogService.destroy();
+      this.contentSearchDialogService.destroy();
+    }
   }
 
   get mode(): ViewerMode {
@@ -146,6 +185,8 @@ export class ViewerComponent implements OnInit, OnDestroy, OnChanges {
     this.viewerService.destroy();
     this.attributionDialogService.destroy();
     this.contentsDialogService.destroy();
+    this.contentSearchDialogService.destroy();
+    this.iiifContentSearchService.destroy();
   }
 
   setClasses() {
